Fix backdrop-filter bug
Chrome has a bug that meant that backdrop-filter didn't work on the header menu because the menu itself was nested.

This fixes the bug by applying the backdrop-filter to a 100% width/height, absolutely positioned :before element on the parent element instead of applying it directly to the parent element. The nested backdrop-filter then works.